day_5

+++ week_11/day_4/Discovering_Permutations.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 typedef long long int ll;
 int main(){
-    int t;
+    ll t;
     cin>>t;
     string s = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
